+++ packages/landingpage-new/.env.development
@@ -1,3 +1,4 @@
 BASE_PATH=landingpage-new
-HOST=http://localhost:4600/$BASE_PATH
+ORIGIN=http://localhost:4600
+HOST=$ORIGIN/$BASE_PATH
 NEXT_PUBLIC_STORYBOOK_URL=http://localhost:6006/
