+++ packages/landingpage-new/.env.production
@@ -1,3 +1,4 @@
 BASE_PATH=landingpage-new
-HOST=https://elements.inovex.de/$BASE_PATH
+ORIGIN=https://elements.inovex.de
+HOST=$ORIGIN/$BASE_PATH
 NEXT_PUBLIC_STORYBOOK_URL=https://elements.inovex.de/version/latest/
