+++ packages/landingpage-new/utils/in-dev-mode.ts
@@ -1,0 +1,7 @@
+let inDevEnvironment = false;
+
+if (process && process.env.NODE_ENV === 'development') {
+  inDevEnvironment = true;
+}
+
+export {inDevEnvironment};
